working on fdw

+++ LdapFdwPlanState.h
@@ -7,7 +7,7 @@
 typedef struct LdapFdwPlanState
 {
 	BlockNumber pages;          /* estimate of file's physical size */
-	double      ntuples;        /* estimate of number of data rows  */
+	int         ntuples;        /* estimate of number of data rows  */
 	List	   *local_conds;
 	List	   *remote_conds;
 } LdapFdwPlanState;
